flask_app/requirements.txt
emoji==2.15.0
Flask==3.1.2
joblib==1.5.2
mlflow==3.6.0
mlflow_skinny==3.6.0
mlflow_tracing==3.6.0
nltk==3.9.2
pandas==2.3.3
PyYAML==6.0.3
scikit_learn==1.7.2
textblob==0.19.0
gunicorn==23.0.0
awscli
boto3